typecast int for length

+++ dags/zetatango_import.py
@@ -171,7 +171,7 @@ def stage_table_in_snowflake(
             if table == "lending_adjudications":
                 lending_csv = csv_filedesc.readlines()
                 len_of_file = sum(1 for row in lending_csv)
-                linesPerFile = len_of_file / 2
+                linesPerFile = int(len_of_file / 2)
                 with open(f"{csv_filepath_split_1}", "w+") as f:
                     for i in range(0, linesPerFile):
                         f.writelines(lending_csv[i])
